Refactoring
Refactoring the game code to seperate all openGL functions from the
main.cpp
