+++ CarreGameEngine/CarreGameEngine/CarreGameEngine.vcxproj
@@ -128,9 +128,9 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="headers\Camera.h" />
+    <ClInclude Include="headers\GameWorld.h" />
     <ClInclude Include="headers\GLFWManager.h" />
     <ClInclude Include="headers\InputManager.h" />
-    <ClInclude Include="headers\Main.h" />
     <ClInclude Include="headers\Model.h" />
     <ClInclude Include="headers\ObjLoader.h" />
     <ClInclude Include="headers\PhysicsEngine.h" />
@@ -148,6 +148,7 @@
     <ClCompile Include="resource\Resource.cpp" />
     <ClCompile Include="resource\ResourceFactory.cpp" />
     <ClCompile Include="src\Camera.cpp" />
+    <ClCompile Include="src\GameWorld.cpp" />
     <ClCompile Include="src\GLFWManager.cpp" />
     <ClCompile Include="src\InputManager.cpp" />
     <ClCompile Include="src\Main.cpp" />
